Add prompts to nag user to make backups whenever the keystore is modified
The following prompt is presented:

```
? The keystore has been updated, have you made a backup of <cfgDir> ?  [Use arrows to move, type to filter]
> Yes, I made a backup
  No, I did not make a backup, abort
  No, I did not make a backup, continue anyways (dangerous!!!)
  How do I make a backup?
```

A new `backups.toml` file is used to record if user has claimed a backup
exists. If so, the nag message won't appear on subsequent cli
invocations.

If new accounts are created, or passphrases changed, the nag message
will re-appear.

The GLIF_BACKUP_EXIST env var can be set to suppress the interactive prompt.