Add --fund-glf-vault and --personal-cashback-percent flags to (#177)
* Add --fund-glf-vault and --personal-cashback-percent flags to
mint-and-activate

Also remove the mint-activate-and-fund command.

One problem I discovered, the SPPlusMintActivateAndFund() method
will revert if the fund amount is zero, so we have to fallback
to the SPPlusMintAndActivate() method, but that doesn't accept
a personal cashback percent param.

For now, I'm just printing out an additional message to the user
in that case:

"Cannot set personal cashback percent in same transaction, please run:
glif plus set-personal-cashback-percent 5.000000"

* Remove log

* Polish

---------

Co-authored-by: Jonathan <jpschwartz2@uwalumni.com>